New implementation of intersection check
The new implementation of checking for segment intersections is more intuitive than the old one. The custom implementation of the cross product is faster than the numpy version, as it can assume the vectors to be always two-dimensional. The cross product for two dimensions corresponds to the magnitude of the 3D cross product, which refers to the 2D matrix determinant.

The new implementation is based on the explanations of @gareth-rees (https://stackoverflow.com/questions/563198/how-do-you-detect-where-two-line-segments-intersect).